Prioritization
This seems like a minor change, but is a big QoL for the players. Previously
the "wrapup" block for each component would preempt even the counters for
the other components, e.g. if you hit the threshold for books and earned a
point then the counters for, say, party diversity would never even get a
chance to increment.

Anything which gives out lore is now pushed to the bottom with a new function
that will keep everything ordered properly. It is a little heavy, code-wise,
but will yield a better player experience.