Different dream script handling (big refactoring)
Basically the dream script was never meant to handle the load of checks we were throwing at it, especially when most of them had
Continue()s thrown in. It was to the point where players thought a rest had been skipped since it was taking so long to process through to
the RestParty() call.

Instead we're spawning an invisible helper creature to handle the scripting instead. While it remains slow--feedback about books read, the
lowest-priority item, seems to take about 5-6 seconds to process and provide feedback--it now runs after RestParty() and is substantially
more reliable.